Move pod yaml viewer related components to separate file.
